requirements.txt
attrs==23.2.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
greenlet==3.0.3
h11==0.14.0
idna==3.7
iniconfig==2.0.0
jedi==0.17.2
outcome==1.3.0.post0
packaging==24.1
parso==0.7.1
pip3-autoremove==1.2.2
playwright==1.47.0
pluggy==1.5.0
pycparser==2.22
pyee==12.0.0
pytest==8.3.1
pytest-asyncio==0.23.8
pytest-subtests==0.13.1
python-dotenv==1.0.1
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
PyYAML==6.0.2
requests==2.32.3
selenium==4.23.0
setuptools==71.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.9.0
ujson==5.10.0
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
